refactor: change rolelisting to roles
