+++ src/i18n/locales/en.json
@@ -8,7 +8,7 @@
   "users": "Users",
   "home": "Home",
   "page": "Page",
-  "roleListing": "Role Listing",
+  "roles": "Role Listing",
   "all": "All",
   "status": "Status",
   "waiting": "Waiting",
