+++ src/i18n/locales/tr.json
@@ -8,7 +8,7 @@
   "users": "Kullanıcılar",
   "home": "Ana Sayfa",
   "page": "Sayfa",
-  "roleListing": "Rol Listeleme",
+  "roles": "Rol Listeleme",
   "all": "Hepsini Seç",
   "status": "Durum",
   "waiting": "Beklemede",
